fix scalar info completion
