+++ package.json
@@ -11,7 +11,9 @@
   "bin": {
     "graphql-binding": "./dist/bin.js"
   },
-  "files": ["dist"],
+  "files": [
+    "dist"
+  ],
   "main": "dist/index.js",
   "typings": "dist/index.d.ts",
   "typescript": {
@@ -20,16 +22,14 @@
   "scripts": {
     "prepublish": "npm run build",
     "build": "rm -rf dist && tsc -d",
-    "test-only":
-      "npm run build && ava --verbose dist/**/*.test.js dist/*.test.js",
-    "test": "tslint src/**/*.ts && npm run test-only",
-    "ava": "ava"
+    "test-only": "npm run build && ava --verbose dist/**/*.test.js dist/*.test.js",
+    "test": "tslint src/**/*.ts && npm run test-only"
   },
   "devDependencies": {
     "@types/graphql": "0.13.0",
     "@types/node": "10.0.2",
     "apollo-link": "1.2.1",
-    "ava": "0.25.0",
+    "ava": "^0.25.0",
     "graphql": "0.13.2",
     "graphql-tag": "^2.9.2",
     "tslint": "5.9.1",
